requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.1.0
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-annoying==0.10.5
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==2.2.5
djangorestframework==3.10.3
freezegun==0.3.12
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-bigquery==1.20.0
google-cloud-core==1.0.3
google-resumable-media==0.4.1
googleads==21.0.0
googleapis-common-protos==1.6.0
httplib2==0.14.0
idna==2.8
isodate==0.6.0
lxml==4.4.1
oauth2client==4.1.3
oauthlib==3.1.0
parameterized==0.7.0
protobuf==3.9.2
psycopg2-binary==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.8.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
xmltodict==0.12.0
zeep==3.4.0
